+++ src/main/java/zighang2/zighang/web/controller/AuthController.java
@@ -1,10 +1,9 @@
 package zighang2.zighang.web.controller;
 
-import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import zighang2.zighang.global.payload.ApiResponse;
-import zighang2.zighang.web.dto.TokenResponse;
+import zighang2.zighang.web.dto.TokenResponseDto;
 import zighang2.zighang.web.service.AuthService;
 
 @RestController
@@ -15,7 +14,7 @@ public class AuthController {
     private final AuthService authService;
 
     @GetMapping("/auth/login/kakao")
-    public ApiResponse<TokenResponse> kakaoLogin(@RequestParam("code") String accessCode) {
+    public ApiResponse<TokenResponseDto> kakaoLogin(@RequestParam("code") String accessCode) {
         return ApiResponse.onSuccess(authService.oAuthLogin(accessCode));
     }
